frontend/requirements.txt
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-bootstrap-templates==1.2.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.6.1
Flask==3.0.3
Flask-Session==0.8.0
httpx==0.27.2
idna==3.7
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
PyJWT==2.9.0
MarkupSafe==2.1.5
msgspec==0.18.6
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
plotly==5.23.0
pycparser==2.22
pymongo==4.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
setuptools==73.0.1
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.20.0
